Add debug switch to disable slow queries such as `keys`.  add GET param `from` and `to` to determin range. and many many enhancement....

+++ flush.php
@@ -8,6 +8,10 @@ if (!isset($_POST['post'])) {
 
 require_once 'includes/common.inc.php';
 
+if ($server['debug'] == true) {
+    $redis->flushdb();
+else {
+    die ("Forbidden operation!");
+}
 
-$redis->flushdb();
